# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MangaGramm API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /series:
    get:
      operationId: listSeries
      tags: [series]
      summary: List all series
      parameters:
        - name: genre
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [ongoing, completed, hiatus]
        - name: type
          in: query
          schema:
            type: string
            enum: [manga, webtoon, comic]
        - name: sort
          in: query
          schema:
            type: string
            enum: [latest, popular, trending]
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeriesListResponse"
    post:
      operationId: createSeries
      tags: [series]
      summary: Create a new series
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSeriesBody"
      responses:
        "201":
          description: Series created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Series"
  /series/{seriesId}:
    get:
      operationId: getSeries
      tags: [series]
      summary: Get series detail
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Series detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeriesDetail"
    put:
      operationId: updateSeries
      tags: [series]
      summary: Update a series
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSeriesBody"
      responses:
        "200":
          description: Series updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Series"
    delete:
      operationId: deleteSeries
      tags: [series]
      summary: Delete a series
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Series deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get user settings
      responses:
        "200":
          description: User settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update user settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsBody"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get user notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
...
  /notifications/{notificationId}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /reading-history/{chapterId}:
    post:
      operationId: trackReading
      tags: [users]
      summary: Track chapter read
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reading tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Series:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        coverImage:
          type: string
          nullable: true
        bannerImage:
          type: string
          nullable: true
        type:
          type: string
          enum: [manga, webtoon, comic]
        status:
          type: string
          enum: [ongoing, completed, hiatus]
        genres:
          type: array
          items:
            type: string
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatar:
          type: string
          nullable: true
        chapterCount:
          type: integer
        viewCount:
          type: integer
        likeCount:
          type: integer
        favoriteCount:
          type: integer
        mature:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - slug
        - type
        - status
        - authorId
        - authorName
        - chapterCount
        - viewCount
        - likeCount
        - createdAt
        - updatedAt
    SeriesDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        coverImage:
          type: string
          nullable: true
        bannerImage:
          type: string
          nullable: true
        type:
          type: string
          enum: [manga, webtoon, comic]
        status:
          type: string
          enum: [ongoing, completed, hiatus]
        genres:
          type: array
          items:
            type: string
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatar:
          type: string
          nullable: true
        chapterCount:
          type: integer
        viewCount:
          type: integer
        likeCount:
          type: integer
        favoriteCount:
          type: integer
        mature:
          type: boolean
        chapters:
          type: array
          items:
            $ref: "#/components/schemas/Chapter"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - slug
        - type
        - status
        - authorId
        - authorName
        - chapterCount
        - viewCount
        - likeCount
        - chapters
        - createdAt
        - updatedAt
    SeriesListResponse:
      type: object
      properties:
        series:
          type: array
          items:
            $ref: "#/components/schemas/Series"
        total:
          type: integer
        hasMore:
          type: boolean
      required:
        - series
        - total
        - hasMore
    CreateSeriesBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        coverImage:
          type: string
        bannerImage:
          type: string
        type:
          type: string
          enum: [manga, webtoon, comic]
        status:
          type: string
          enum: [ongoing, completed, hiatus]
          default: ongoing
        genres:
          type: array
          items:
            type: string
        mature:
          type: boolean
          default: false
      required:
        - title
        - type
    UpdateSeriesBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        coverImage:
          type: string
        bannerImage:
          type: string
        status:
          type: string
          enum: [ongoing, completed, hiatus]
        genres:
          type: array
          items:
            type: string
        mature:
          type: boolean
    Chapter:
      type: object
      properties:
        id:
          type: integer
        seriesId:
          type: integer
        number:
          type: number
        title:
          type: string
        pageCount:
          type: integer
        viewCount:
          type: integer
        likeCount:
          type: integer
        commentCount:
          type: integer
        publishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - seriesId
        - number
        - title
        - pageCount
        - viewCount
        - likeCount
        - commentCount
        - createdAt
...
    UserSettings:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        theme:
          type: string
          enum: [light, dark, system]
        language:
          type: string
        emailNotifications:
          type: boolean
        pushNotifications:
          type: boolean
        matureContent:
          type: boolean
        readingDirection:
          type: string
          enum: [ltr, rtl]
        autoNextChapter:
          type: boolean
        pageLayout:
          type: string
          enum: [single, double, scroll]
      required:
        - userId
        - theme
        - language
        - emailNotifications
        - pushNotifications
        - matureContent
        - readingDirection
        - autoNextChapter
        - pageLayout
    UpdateSettingsBody:
      type: object
      properties:
        theme:
          type: string
          enum: [light, dark, system]
        language:
          type: string
        emailNotifications:
          type: boolean
        pushNotifications:
          type: boolean
        matureContent:
          type: boolean
        readingDirection:
          type: string
          enum: [ltr, rtl]
        autoNextChapter:
          type: boolean
        pageLayout:
          type: string
          enum: [single, double, scroll]
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [new_chapter, new_follower, like, comment, system]
        title:
          type: string
        message:
          type: string
        link:
          type: string
          nullable: true
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - title
        - message
        - read
        - createdAt
...